refactor: reorganize project structure and remove unused files
- Restructured pages into dedicated directories
- Removed deprecated test configurations
- Cleaned up unused component files
- Consolidated portal-related components
- Updated routing structure
