--- jest.config.js
@@ -1,29 +1,0 @@
-export default {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  roots: ['<rootDir>/tests', '<rootDir>/src'],
-  moduleNameMapper: {
-    '\\.(css|less|scss)$': 'identity-obj-proxy',
-    '^@/(.*)$': '<rootDir>/src/$1'
-  },
-  setupFilesAfterEnv: ['<rootDir>/tests/setup.ts'],
-  testMatch: [
-    '**/tests/**/*.test.[jt]s?(x)',
-    '**/tests/integration/**/*.test.[jt]s?(x)',
-    '**/tests/e2e/**/*.spec.[jt]s'
-  ],
-  collectCoverageFrom: [
-    'src/**/*.{ts,tsx}',
-    '!src/**/*.d.ts',
-    '!src/vite-env.d.ts',
-    '!src/main.tsx'
-  ],
-  coverageThreshold: {
-    global: {
-      branches: 80,
-      functions: 80,
-      lines: 80,
-      statements: 80
-    }
-  }
-};
