--- src/features/auth/hooks/useRegisterForm.ts
@@ -1,89 +1,0 @@
-import { useState } from 'react';
-import { useAuth } from '../context/AuthContext';
-import { useNavigate } from 'react-router-dom';
-import { registerSchema } from '../schemas/auth.schema';
-import { z } from 'zod';
-
-interface RegisterForm {
-  email: string;
-  password: string;
-  confirmPassword: string;
-  firstName: string;
-  lastName: string;
-  phoneNumber: string;
-  address: string;
-}
-
-type RegisterFormErrors = z.inferFormattedError<typeof registerSchema>;
-
-export const useRegisterForm = () => {
-  const [formData, setFormData] = useState<RegisterForm>({
-    email: '',
-    password: '',
-    confirmPassword: '',
-    firstName: '',
-    lastName: '',
-    phoneNumber: '',
-    address: '',
-  });
-  const [errors, setErrors] = useState<RegisterFormErrors['fieldErrors']>({});
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const { register } = useAuth();
-  const navigate = useNavigate();
-
-  const validateForm = (): boolean => {
-    try {
-      registerSchema.parse(formData);
-      setErrors({});
-      return true;
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        setErrors(error.format().fieldErrors);
-      }
-      return false;
-    }
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({ ...prev, [name]: value }));
-    
-    if (errors[name as keyof RegisterForm]) {
-      setErrors((prev) => ({ ...prev, [name]: undefined }));
-    }
-  };
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    
-    if (!validateForm()) {
-      return;
-    }
-    
-    setIsSubmitting(true);
-    
-    try {
-      await register(
-        formData.email,
-        formData.password,
-        formData.firstName,
-        formData.lastName,
-        formData.phoneNumber,
-        formData.address
-      );
-      navigate('/portal');
-    } catch (error) {
-      console.error('Registration failed:', error);
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  return {
-    formData,
-    errors,
-    isSubmitting,
-    handleInputChange,
-    handleSubmit,
-  };
-};
