--- src/pages/EmployeePortalPage.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import EmployeePortal from '../features/employeePortal/components/EmployeePortal';
-
-const EmployeePortalPage: React.FC = () => {
-  return <EmployeePortal />;
-};
-
-export default EmployeePortalPage;
