--- tests/setup.ts
@@ -1,45 +1,0 @@
-// tests/setup.ts
-import '@testing-library/jest-dom';
-
-// Commenting out MSW until it's installed
-// import { setupServer } from 'msw/node';
-// import { handlers } from './mocks/handlers';
-
-// Setup MSW
-// export const server = setupServer(...handlers);
-
-// beforeAll(() => server.listen());
-// afterEach(() => server.resetHandlers());
-// afterAll(() => server.close());
-
-// Mock window.matchMedia
-Object.defineProperty(window, 'matchMedia', {
-  writable: true,
-  value: jest.fn().mockImplementation(query => ({
-    matches: false,
-    media: query,
-    onchange: null,
-    addListener: jest.fn(),
-    removeListener: jest.fn(),
-    addEventListener: jest.fn(),
-    removeEventListener: jest.fn(),
-    dispatchEvent: jest.fn(),
-  })),
-});
-
-// Mock IntersectionObserver
-class MockIntersectionObserver {
-  observe = jest.fn();
-  unobserve = jest.fn();
-  disconnect = jest.fn();
-}
-
-Object.defineProperty(window, 'IntersectionObserver', {
-  writable: true,
-  value: MockIntersectionObserver,
-});
-
-// Clean up after each test
-afterEach(() => {
-  jest.clearAllMocks();
-});
