--- tests/unit/shared/components/Button.test.tsx
@@ -1,45 +1,0 @@
-import React from 'react';
-import { render, screen, fireEvent } from '../../utils/render';
-import Button from '../../../../src/shared/components/Button';
-
-describe('Button', () => {
-  it('renders children correctly', () => {
-    render(<Button>Click me</Button>);
-    expect(screen.getByText('Click me')).toBeInTheDocument();
-  });
-
-  it('handles click events', () => {
-    const handleClick = jest.fn();
-    render(<Button onClick={handleClick}>Click me</Button>);
-    
-    fireEvent.click(screen.getByText('Click me'));
-    expect(handleClick).toHaveBeenCalledTimes(1);
-  });
-
-  it('applies variant classes correctly', () => {
-    const { rerender } = render(<Button variant="primary">Primary</Button>);
-    expect(screen.getByText('Primary')).toHaveClass('bg-primary-600');
-
-    rerender(<Button variant="secondary">Secondary</Button>);
-    expect(screen.getByText('Secondary')).toHaveClass('bg-gray-100');
-  });
-
-  it('shows loading state', () => {
-    render(<Button isLoading>Loading</Button>);
-    expect(screen.getByRole('button')).toBeDisabled();
-    expect(screen.getByText('Loading')).toBeInTheDocument();
-  });
-
-  it('applies full width class when fullWidth prop is true', () => {
-    render(<Button fullWidth>Full Width</Button>);
-    expect(screen.getByRole('button')).toHaveClass('w-full');
-  });
-
-  it('renders with icon', () => {
-    const TestIcon = () => <svg data-testid="test-icon" />;
-    render(<Button icon={<TestIcon />}>With Icon</Button>);
-    
-    expect(screen.getByTestId('test-icon')).toBeInTheDocument();
-    expect(screen.getByText('With Icon')).toBeInTheDocument();
-  });
-});
